app/backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
SQLAlchemy==2.0.28
psycopg2-binary==2.9.9
alembic==1.13.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.6.3
python-multipart==0.0.9
requests==2.31.0
